requirements.txt
crewai
python-dotenv
discord.py